[REF] move mwps to medor_custom
