+++ medor_custom/controllers/__init__.py
@@ -2,3 +2,4 @@ from . import trial_subscription
 from . import delivery_form
 from . import delivery
 from . import subscribe
+from . import user
