+++ medor_website_product_subscription/__openerp__.py
@@ -3,10 +3,9 @@
 #     Rémy Taymans <remy@coopiteasy.be>
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 {
-    # todo move to medor_custom
     "name": "Medor Website Product Subscription",
     "summary": """
-        Add Medor specific change to Website Product Subscription.
+        empty module - can be uninstalled
     """,
     "description": """
     """,
@@ -15,6 +14,6 @@
     "license": "AGPL-3",
     "category": "Website",
     "version": "9.0.1.0",
-    "depends": ["website_product_subscription"],
-    "data": ["views/medor_website_ps_templates.xml"],
+    "depends": [],
+    "data": [],
 }
