--- medor_website_product_subscription/controllers/__init__.py
@@ -1,1 +1,0 @@
-from . import main
